+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 - raised esbuild target to `esnext`
 - added `make()` to make a game object without adding to the scene
 - `add()` supports adding unattached game objects to the scene `add(make([...comps]))`
+- added `colors` option to `drawPolygon()`
+- added `gradient` option to `drawRect()`
 
 ### v2000.2.2